requirements.txt
langchain>=0.3.0
langchain-openai>=0.1.0
langgraph>=0.2.30
pymilvus>=2.4.5
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0
